+++ packages/api/lib/testUtils.js
@@ -27,7 +27,6 @@ const {
   indexRule,
   indexPdr,
   indexAsyncOperation,
-  indexExecution,
   deleteExecution,
 } = require('@cumulus/es-client/indexer');
 const {
@@ -610,8 +609,6 @@ const createExecutionTestRecords = async (context, executionParams = {}) => {
   const {
     knex,
     executionPgModel,
-    esClient,
-    esExecutionsClient,
   } = context;
 
   const originalExecution = fakeExecutionFactoryV2(executionParams);
@@ -621,13 +618,8 @@ const createExecutionTestRecords = async (context, executionParams = {}) => {
   const originalPgRecord = await executionPgModel.get(
     knex, { cumulus_id: executionCumulusId }
   );
-  await indexExecution(esClient, originalExecution, process.env.ES_INDEX);
-  const originalEsRecord = await esExecutionsClient.get(
-    originalExecution.arn
-  );
   return {
     originalPgRecord,
-    originalEsRecord,
   };
 };