+++ packages/db/src/search/CollectionSearch.ts
@@ -53,7 +53,7 @@ export class CollectionSearch extends BaseSearch {
       searchQuery: Knex.QueryBuilder,
     } {
     const countQuery = knex(this.tableName)
-      .count(`${this.tableName}.cumulus_id`);
+      .count('*');
 
     const searchQuery = knex(this.tableName)
       .select(`${this.tableName}.*`);
@@ -136,7 +136,7 @@ export class CollectionSearch extends BaseSearch {
     const granulesTable = TableNames.granules;
     const statsQuery = knex(granulesTable)
       .select(`${granulesTable}.collection_cumulus_id`, `${granulesTable}.status`)
-      .count(`${granulesTable}.status`)
+      .count('*')
       .groupBy(`${granulesTable}.collection_cumulus_id`, `${granulesTable}.status`)
       .whereIn(`${granulesTable}.collection_cumulus_id`, collectionCumulusIds);